Introduce a testing generator
This introduces `suspenders:testing`, a generator that installs our
testing infrastructure. Specifically, rspec and shoulda-matchers.

The `#empty_directory_with_keep_file` method is defined in
`Rails::Generators::AppBase`, but that is too heavy of a dependency;
instead, define the method manually.
