+++ lib/suspenders.rb
@@ -12,6 +12,7 @@ require "suspenders/generators/jobs_generator"
 require "suspenders/generators/analytics_generator"
 require "suspenders/generators/views_generator"
 require "suspenders/generators/js_driver_generator"
+require "suspenders/generators/testing_generator"
 require "suspenders/actions"
 require "suspenders/adapters/heroku"
 require "suspenders/app_builder"
