+++ lib/suspenders/app_builder.rb
@@ -176,29 +176,6 @@ config.public_file_server.headers = {
       create_file '.ruby-version', "#{Suspenders::RUBY_VERSION}\n"
     end
 
-    def provide_shoulda_matchers_config
-      copy_file(
-        "shoulda_matchers_config_rspec.rb",
-        "spec/support/shoulda_matchers.rb"
-      )
-    end
-
-    def configure_spec_support_features
-      empty_directory_with_keep_file 'spec/features'
-      empty_directory_with_keep_file 'spec/support/features'
-    end
-
-    def configure_rspec
-      remove_file "spec/rails_helper.rb"
-      remove_file "spec/spec_helper.rb"
-      copy_file "rails_helper.rb", "spec/rails_helper.rb"
-      copy_file "spec_helper.rb", "spec/spec_helper.rb"
-    end
-
-    def configure_i18n_for_test_environment
-      copy_file "i18n.rb", "spec/support/i18n.rb"
-    end
-
     def configure_i18n_for_missing_translations
       raise_on_missing_translations_in("development")
       raise_on_missing_translations_in("test")
@@ -231,10 +208,6 @@ Rack::Timeout.timeout = (ENV["RACK_TIMEOUT"] || 10).to_i
         "production",
         %{ENV.fetch("ASSET_HOST", ENV.fetch("APPLICATION_HOST"))},
       )
-    end
-
-    def generate_rspec
-      generate 'rspec:install'
     end
 
     def replace_default_puma_configuration
